Move pagination from profile to profile_sections for better readability
